feat(profile) - fixed profile screen location
