+++ presentation/src/main/java/com/moksh/presentation/ui/common/Extensions.kt
@@ -1,0 +1,18 @@
+package com.moksh.presentation.ui.common
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
+
+@Composable
+fun Modifier.clickable(enabled: Boolean = true, onClick: () -> Unit): Modifier = composed {
+    clickable(
+        enabled = enabled,
+        onClick = onClick,
+        indication = null,
+        interactionSource = remember { MutableInteractionSource() }
+    )
+}
