+++ presentation/src/main/java/com/moksh/presentation/ui/profile/components/ProfileEditTextField.kt
@@ -1,4 +1,4 @@
-package com.moksh.presentation.ui.profile_tab.components
+package com.moksh.presentation.ui.profile.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
