--- presentation/src/main/java/com/moksh/presentation/ui/profile_tab/components/ProfileDataBox.kt
@@ -1,37 +1,0 @@
-package com.moksh.presentation.ui.profile_tab.components
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.moksh.presentation.ui.common.Gap
-
-@Composable
-fun ProfileDataBox(
-    modifier: Modifier = Modifier,
-    heading: String,
-    content: String,
-) {
-    Column(
-        modifier = modifier.fillMaxWidth(),
-        horizontalAlignment = Alignment.Start
-    ) {
-        Text(
-            text = heading,
-            style = MaterialTheme.typography.bodyMedium.copy(
-                color = MaterialTheme.colorScheme.onBackground.copy(.7f)
-            )
-        )
-        Gap(size = 6.dp)
-        Text(
-            text = content,
-            style = MaterialTheme.typography.bodyLarge.copy(
-                color = MaterialTheme.colorScheme.onBackground
-            )
-        )
-    }
-}
